feat(gh-actions): Declare reusable workflow inputs and outputs
By decalring outputs from the ReleaseAndBuild workflow, it's possible to
pass them to subsequent workflows which depend on, or "needs",
ReleaseAndBuild.  In this case, ReleaseAndBuild outputs whether or not a
new release was published (as a string, "true" or "false") and the
release version (e.g. "v1.2.3").  The Helm workflow takes as input these
same values and will be able to use them instead of the extra steps
currently required to derive those facts.